requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
mongoengine>=0.27
pymongo>=3.12,<5.0
channels>=4.0
channels-redis>=4.0
python-dotenv>=1.0
PyJWT>=2.8
redis>=4.0
